+++ src/components/UserDetailSections/ReadingRoomAccess/ReadingRoomAccess.js
@@ -1,8 +1,8 @@
-import React, { useState } from 'react';
+import React, { useState, isValidElement, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage, useIntl } from 'react-intl';
-// import { Link } from 'react-router-dom';
-// import { get } from 'lodash';
+import { Link } from 'react-router-dom';
+import { get } from 'lodash';
 
 import {
   Accordion,
@@ -10,111 +10,100 @@ import {
   FormattedUTCDate,
   Headline,
   MultiColumnList,
+  NoValue,
   SearchField,
+  Row,
+  Column,
 } from '@folio/stripes/components';
-// import { ViewMetaData } from '@folio/stripes/smart-components';
-import { READING_ROOM_ACCESS } from '../../../constants';
-// import css from './ReadingRoom.css';
+import { ViewMetaData } from '@folio/stripes/smart-components';
 
-const mockedRRAData = [
-  {
-    'id': 'id1',
-    'userId': 'userId1',
-    'readingRoomId': 'rrId1',
-    'readingRoomName': 'reading room 1',
-    'access': 'NOT_ALLOWED',
-    'notes': 'string',
-    'metadata': {
-      'createdDate': 'string',
-      'createdByUserId': 'c569031b-2399-563c-a8ab-20a2d2a47ba2',
-      'updatedDate': '2024-04-03T02:06:36.761+00:00',
-      'updatedByUserId': 'b17096d8-5b8c-587f-8e78-8f808994b09b',
-    }
-  },
-  {
-    'id': 'id2',
-    'userId': 'userId2',
-    'readingRoomId': 'rrId2',
-    'readingRoomName': 'reading room 2',
-    'access': 'ALLOWED',
-    'notes': 'string',
-    'metadata': {
-      'createdDate': 'string',
-      'createdByUserId': 'c569031b-2399-563c-a8ab-20a2d2a47ba2',
-      'updatedDate': '2024-04-03T02:06:36.761+00:00',
-      'updatedByUserId': 'b17096d8-5b8c-587f-8e78-8f808994b09b',
-    }
-  }
-];
+import { READING_ROOM_ACCESS } from '../../../constants';
+import { rraColumns } from './constant';
+import css from './ReadingRoomAccess.css';
 
 const ReadingRoomAccess = (props) => {
-  const intl = useIntl();
-  const [filteredRRA, setFilteredRRA] = useState(mockedRRAData);
-
   const {
     accordionId,
     expanded,
     onToggle,
-    // userId,
+    userRRAPermissions,
   } = props;
+  const intl = useIntl();
+  const [filteredRRA, setFilteredRRA] = useState([]);
+
+  useEffect(() => {
+    setFilteredRRA(userRRAPermissions);
+  }, [userRRAPermissions]);
 
   const filterReadingRoomsByName = (e) => {
     const name = e.target.value;
-    const filteredRRs = mockedRRAData.filter(r => r.readingRoomName.includes(name));
+    const filteredRRs = userRRAPermissions.filter(r => r.readingRoomName.includes(name));
     setFilteredRRA(filteredRRs);
   };
 
-  // const renderName = (usr) => {
-  //   const lastName = get(usr, ['personal', 'lastName'], '');
-  //   const firstName = get(usr, ['personal', 'firstName'], '');
-  //   const middleName = get(usr, ['personal', 'middleName'], '');
+  const renderName = (usr) => {
+    const lastName = get(usr, ['personal', 'lastName'], '');
+    const firstName = get(usr, ['personal', 'firstName'], '');
+    const middleName = get(usr, ['personal', 'middleName'], '');
 
-  //   return `${lastName}${firstName ? ', ' : ' '}${firstName} ${middleName}`;
-  // };
+    return `${lastName}${firstName ? ', ' : ' '}${firstName} ${middleName}`;
+  };
 
-  // const renderUser = (usr) => {
-  //   if (typeof usr === 'string' || isValidElement(usr)) return usr;
+  const renderUser = (usr) => {
+    if (typeof usr === 'string' || isValidElement(usr)) return usr;
 
-  //   const name = renderName(usr);
+    const name = renderName(usr);
 
-  //   if (usr?.id) {
-  //     return <Link to={`/users/view/${usr.id}`} data-test-user-link>{name}</Link>;
-  //   } else {
-  //     return name;
-  //   }
-  // };
+    if (usr?.id) {
+      return <Link to={`/users/view/${usr.id}`} data-test-user-link>{name}</Link>;
+    } else {
+      return name;
+    }
+  };
 
-  // const lastUpdatedDetails = (updater, date) => {
-  //   return (
-  //     <div className={css.lastUpdatedBy}>
-  //       <FormattedUTCDate value={date} />
-  //       {renderUser(updater, date) }
-  //     </div>
-  //   );
-  // };
+  const lastUpdatedDetails = (updater, date) => {
+    const lastUpdatedUser = renderUser(updater);
+    return (
+      <div className={css.lastUpdatedBy}>
+        <FormattedMessage
+          id="ui-users.reading-room-access.metaSection.lastUpdatedDetails"
+          values={{
+            date: <FormattedUTCDate value={date} />,
+            name: lastUpdatedUser,
+          }}
+        />
+      </div>
+    );
+  };
 
-  const visibleColumns = ['access', 'readingRoomName', 'notes', 'id'];
+  const visibleColumns = [
+    rraColumns.ACCESS,
+    rraColumns.READING_ROOM_NAME,
+    rraColumns.NOTES,
+    rraColumns.ID
+  ];
   const columnMapping = {
-    access: <FormattedMessage id="ui-users.readingRoom.access" />,
-    readingRoomName: <FormattedMessage id="ui-users.readingRoom.name" />,
-    notes: <FormattedMessage id="ui-users.readingRoom.note" />,
-    id: <FormattedMessage id="ui-users.readingRoom.lastUpdated" />,
+    [rraColumns.ACCESS]: <FormattedMessage id="ui-users.readingRoom.access" />,
+    [rraColumns.READING_ROOM_NAME]: <FormattedMessage id="ui-users.readingRoom.name" />,
+    [rraColumns.NOTES]: <FormattedMessage id="ui-users.readingRoom.note" />,
+    [rraColumns.ID]: <FormattedMessage id="ui-users.readingRoom.lastUpdated" />,
   };
   const formatter = {
-    access: (rra) => <FormattedMessage id={READING_ROOM_ACCESS[rra.access]} />,
-    readingRoomName: (rra) => rra.readingRoomName,
-    notes: rra => rra.notes,
-    // id: rra => (
-    //   <ViewMetaData
-    //     metadata={rra.metadata}
-    //   >
-    //     {
-    //       (updater) => {
-    //         return lastUpdatedDetails(updater?.lastUpdatedBy, rra.metadata.updatedDate);
-    //       }}
-    //   </ViewMetaData>
-    // )
-    id: rra => <FormattedUTCDate value={rra.updatedDate} />
+    [rraColumns.ACCESS]: (rra) => <FormattedMessage id={READING_ROOM_ACCESS[rra.access]} />,
+    [rraColumns.READING_ROOM_NAME]: (rra) => rra.readingRoomName,
+    [rraColumns.NOTES]: rra => rra.notes,
+    [rraColumns.ID]: rra => (
+      rra?.metadata?.updatedDate ? (
+        <ViewMetaData
+          metadata={rra.metadata}
+        >
+          {
+          (updater) => lastUpdatedDetails(updater?.lastUpdatedBy, rra.metadata.updatedDate)
+        }
+        </ViewMetaData>
+      ) :
+        <NoValue />
+    )
   };
 
   return (
@@ -127,11 +116,16 @@ const ReadingRoomAccess = (props) => {
         <Badge>{filteredRRA.length}</Badge>
       }
       displayWhenOpen={
-        <SearchField
-          onChange={filterReadingRoomsByName}
-          onClear={() => setFilteredRRA(mockedRRAData)}
-          placeholder={intl.formatMessage({ id:'ui-users.readingRoom.filter' })}
-        />
+        <div
+          className={css.rraSearchFieldContainer}
+        >
+          <SearchField
+            onChange={filterReadingRoomsByName}
+            onClear={() => setFilteredRRA(userRRAPermissions)}
+            placeholder={intl.formatMessage({ id:'ui-users.readingRoom.filter' })}
+          />
+        </div>
+        // <p>Add item</p>
       }
     >
       <MultiColumnList
@@ -150,12 +144,7 @@ ReadingRoomAccess.propTypes = {
   expanded: PropTypes.bool,
   onToggle: PropTypes.func,
   accordionId: PropTypes.string.isRequired,
-  // userId: PropTypes.string,
+  userRRAPermissions: PropTypes.arrayOf(PropTypes.object),
 };
-// ReadingRoomAccess.manifest = Object.freeze({
-//   readingRoomAccess: {
-//     type: 'okapi',
-//     path: '/reading-room-patron-permission/{props.userId}'
-//   }
-// });
+
 export default ReadingRoomAccess;