+++ src/components/UserDetailSections/ReadingRoomAccess/ReadingRoomAccess.test.js
@@ -15,11 +15,32 @@ jest.mock('@folio/stripes/components', () => ({
   )),
 }));
 
+const mockedRRAData = [
+  {
+    'userId': '2205005b-ca51-4a04-87fd-938eefa8f6de',
+    'readingRoomId': '0e589a51-5f28-4b59-b923-b036cf56a41d',
+    'readingRoomName': 'reading room1',
+    'access': 'NOT_ALLOWED'
+  },
+  {
+    'id': '2205005b-ca51-4a04-87fd-938eefaf86de',
+    'userId': '2205005b-ca51-4a04-87fd-938eefa8f6de',
+    'readingRoomId': 'ef1b8f7f-afb3-4454-bc3a-4c9989fcf977',
+    'readingRoomName': 'reading room2',
+    'access': 'NOT_ALLOWED',
+    'metadata': {
+      'createdDate': '2024-05-03T09:18:16.221500',
+      'createdByUserId': '21457ab5-4635-4e56-906a-908f05e9233b',
+      'updatedDate': '2024-05-03T09:18:52.016583',
+      'updatedByUserId': '21457ab5-4635-4e56-906a-908f05e9233b'
+    }
+  },
+];
 const props = {
   accordionId: 'readingRoomSection',
   expanded: false,
   onToggle: jest.fn(),
-  // userId: 'b17096d8-5b8c-587f-8e78-8f808994b09b',
+  userRRAPermissions: mockedRRAData,
 };
 
 describe('ReadingRoomAccess', () => {